Centralise adding columns.
The document object no longer exposes a mutable array of columns and requires that anything that wishes to add them do so by calling a method on the document object (removing not yet supported).  A notification is now fired when a column is added, allowing the UI to be updated.